Set ui ordering for things like return_pressed()

+++ Source/Time-Track/mainform.cpp
@@ -1156,7 +1156,7 @@ void MainForm::refreshProjectItemTable(){
         QString a = QString::number(((double)ui->ProjectItemWidget->item(i,7)->text().toInt())/ui->ProjectItemWidget->item(i,3)->text().toInt(),'f',2);
         ui->ProjectItemWidget->setItem(i,8,new QTableWidgetItem(a+" HR/"+ui->ProjectItemWidget->item(i,4)->text()));
 
-        QString p = QString::number(((double)(100*ui->ProjectItemWidget->item(i,7)->text().toInt()))/ui->ProjectItemWidget->item(i,5)->text().toInt(),'f',2);
+        QString p = QString::number(((double)(100*ui->ProjectItemWidget->item(i,7)->text().toDouble()))/ui->ProjectItemWidget->item(i,5)->text().toDouble(),'f',2);
         if( ui->ProjectItemWidget->item(i,5)->text().toInt() == 0){
             ui->ProjectItemWidget->setItem(i,9,new QTableWidgetItem(p));
         }else{
